+++ packages/automations/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @vendoai/automations
 
+## 0.30.1
+
+### Patch Changes
+
+- @vendoai/core@0.30.1
+
+## 0.30.0
+
+### Patch Changes
+
+- Updated dependencies [56c81b5]
+  - @vendoai/core@0.30.0
+
 ## 0.29.1
 
 ### Patch Changes