+++ packages/store/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @vendoai/store
 
+## 0.30.1
+
+### Patch Changes
+
+- Updated dependencies [6bbc8e6]
+  - @vendoai/apps@0.30.1
+  - @vendoai/core@0.30.1
+
+## 0.30.0
+
+### Patch Changes
+
+- Updated dependencies [b3d92b2]
+- Updated dependencies [bd1d016]
+- Updated dependencies [56c81b5]
+  - @vendoai/apps@0.30.0
+  - @vendoai/core@0.30.0
+
 ## 0.29.1
 
 ### Patch Changes