+++ packages/vendoai/CHANGELOG.md
@@ -1,5 +1,18 @@
 # vendoai
 
+## 0.30.1
+
+### Patch Changes
+
+- @vendoai/vendo@0.30.1
+
+## 0.30.0
+
+### Patch Changes
+
+- Updated dependencies [bd1d016]
+  - @vendoai/vendo@0.30.0
+
 ## 0.29.1
 
 ### Patch Changes